+++ src/app/rest.container.ts
@@ -6,8 +6,8 @@ import type { ConfigInterface } from '../core/config/config.interface.js';
 import ConfigService from '../core/config/config.service.js';
 import type { DatabaseClientInterface } from '../core/database-client/mongo-client.interface.js';
 import MongoClientService from '../core/database-client/database-client.service.js';
-import type { ExceptionFilterInterface } from '../core/expception-filters/exception-filter.interface.js';
-import ExceptionFilter from '../core/expception-filters/exception-filter.js';
+import type { ExceptionFilterInterface } from '../core/exception-filters/exception-filter.interface.js';
+import ExceptionFilter from '../core/exception-filters/exception-filter.js';
 import type { LoggerInterface } from '../core/logger/logger.interface.js';
 import PinoService from '../core/logger/pino.service.js';
 import RestApplication from './rest.js';
