+++ src/app/rest.ts
@@ -9,10 +9,9 @@ import type { DatabaseClientInterface } from '../core/database-client/mongo-clie
 import { AppComponent } from '../types/app-component.enum.js';
 import { getMongoURI } from '../core/helpers/db.js';
 import { ControllerInterface } from '../core/controller/controller.interface.js';
-import { ExceptionFilterInterface } from '../core/expception-filters/exception-filter.interface.js';
+import { ExceptionFilterInterface } from '../core/exception-filters/exception-filter.interface.js';
 import { AuthenticateMiddleware } from '../core/middlewares/authenticate.middleware.js';
 import { getFullServerPath } from '../core/helpers/common.js';
-
 
 @injectable()
 export default class RestApplication {
@@ -30,7 +29,7 @@ export default class RestApplication {
     this.expressApplication = express();
   }
 
-  private async _initExceptionFilters() {
+  private async initExceptionFilters() {
     this.logger.info('Exception filters initialization');
 
     this.expressApplication.use(this.exceptionFilter.catch.bind(this.exceptionFilter));
@@ -38,7 +37,7 @@ export default class RestApplication {
     this.logger.info('Exception filters completed');
   }
 
-  private async _initMiddleware() {
+  private async initMiddleware() {
     this.logger.info('Global middleware initialization…');
 
     this.expressApplication.use(express.json());
@@ -58,7 +57,7 @@ export default class RestApplication {
     this.logger.info('Global middleware initialization completed');
   }
 
-  private async _initRoutes() {
+  private async initRoutes() {
     this.logger.info('Controller initialization…');
 
     this.expressApplication.use('/users', this.userController.router);
@@ -68,7 +67,7 @@ export default class RestApplication {
     this.logger.info('Controller initialization completed');
   }
 
-  private async _initDb() {
+  private async initDb() {
     this.logger.info('Init database…');
 
     const mongoUri = getMongoURI(
@@ -84,7 +83,7 @@ export default class RestApplication {
     this.logger.info('Init database completed');
   }
 
-  private async _initServer() {
+  private async initServer() {
     this.logger.info('Try to init server...');
 
     const host = this.config.get('HOST');
@@ -97,17 +96,17 @@ export default class RestApplication {
   public async init() {
     this.logger.info('Application initialization…');
 
-    await this._initDb().catch((error) => {
+    await this.initDb().catch((error) => {
       this.logger.error(`Error during database initialization: ${error.message}`);
     });
 
-    await this._initMiddleware();
+    await this.initMiddleware();
 
-    await this._initRoutes();
+    await this.initRoutes();
 
-    await this._initExceptionFilters();
+    await this.initExceptionFilters();
 
-    await this._initServer().catch((error) => {
+    await this.initServer().catch((error) => {
       this.logger.error(`Error server initialization: ${error.message}`);
     });
   }
