+++ src/core/exception-filters/exception-filter.interface.ts
